+++ Gemfile
@@ -1,8 +1,12 @@
 source 'https://rubygems.org'
 
 group :test do
+	gem 'envyable', '~> 1.2' 
+	gem 'http', '~> 4.4'
   gem 'rspec'
   gem 'rubocop', '0.79.0'
   gem 'simplecov', require: false, group: :test
   gem 'simplecov-console', require: false, group: :test
+	gem 'sinatra'
+	gem 'twilio-ruby'
 end
