--- docs/script.rb
@@ -1,4 +1,0 @@
-require_relative 'takeaway_app'
-require_relative 'dish'
-require_relative 'order'
-
