+++ lib/dish.rb
@@ -1,0 +1,8 @@
+class Dish
+		attr_reader :name, :price 
+		def initialize(name, price)
+				@name = name
+				@price = price
+		end
+end
+
