+++ lib/twilio.rb
@@ -1,9 +1,10 @@
 require 'twilio-ruby'
 require 'sinatra'
 
+=begin
 account_sid = ACCOUNT_SID 
 auth_token = AUTH_TOKEN 
 
 # set up a client to talk to the Twilio REST API
 @clinet = Twilio::REST::Client.new account_sid, auth_token
-
+=end
